Rename before any changes are made so GitHub shows the changes correctly
